Added checking when list changes and no subscriber object is returned

+++ app/code/community/Ebizmarts/MailChimp/Model/ProcessWebhook.php
@@ -78,12 +78,14 @@ class Ebizmarts_MailChimp_Model_ProcessWebhook
         $oldSubscriber = Mage::helper('mailchimp')->loadListSubscriber($listId, $old);
         $newSubscriber = Mage::helper('mailchimp')->loadListSubscriber($listId, $new);
 
-        if (!$newSubscriber->getId()) {
-            if ($oldSubscriber->getId()) {
-                $oldSubscriber->setSubscriberEmail($new)
-                    ->save();
-            } else {
-                $this->subscribeMember($newSubscriber);
+        if ($oldSubscriber) {
+            if (!$newSubscriber->getId()) {
+                if ($oldSubscriber->getId()) {
+                    $oldSubscriber->setSubscriberEmail($new)
+                        ->save();
+                } else {
+                    $this->subscribeMember($newSubscriber);
+                }
             }
         }
     }
@@ -99,7 +101,7 @@ class Ebizmarts_MailChimp_Model_ProcessWebhook
         //Delete subscriber from Magento
         $s = Mage::helper('mailchimp')->loadListSubscriber($data['list_id'], $data['email']);
 
-        if ($s->getId()) {
+        if ($s && $s->getId()) {
             try {
                 $s->delete();
             } catch (Exception $e) {
@@ -120,21 +122,23 @@ class Ebizmarts_MailChimp_Model_ProcessWebhook
             $listId = $data['list_id'];
             $email = $data['email'];
             $subscriber = Mage::helper('mailchimp')->loadListSubscriber($listId, $email);
-            if ($subscriber->getId()) {
-                if ($subscriber->getSubscriberStatus() != Mage_Newsletter_Model_Subscriber::STATUS_SUBSCRIBED) {
+            if ($subscriber) {
+                if ($subscriber->getId()) {
+                    if ($subscriber->getSubscriberStatus() != Mage_Newsletter_Model_Subscriber::STATUS_SUBSCRIBED) {
+                        $this->subscribeMember($subscriber);
+                    }
+                } else {
+                    if (isset($data['merges']['FNAME'])) {
+                        $subscriberFname = filter_var($data['merges']['FNAME'], FILTER_SANITIZE_STRING);
+                        $subscriber->setSubscriberFirstname($subscriberFname);
+                    }
+
+                    if (isset($data['merges']['LNAME'])) {
+                        $subscriberLname = filter_var($data['merges']['LNAME'], FILTER_SANITIZE_STRING);
+                        $subscriber->setSubscriberLastname($subscriberLname);
+                    }
                     $this->subscribeMember($subscriber);
                 }
-            } else {
-                if (isset($data['merges']['FNAME'])) {
-                    $subscriberFname = filter_var($data['merges']['FNAME'], FILTER_SANITIZE_STRING);
-                    $subscriber->setSubscriberFirstname($subscriberFname);
-                }
-
-                if (isset($data['merges']['LNAME'])) {
-                    $subscriberLname = filter_var($data['merges']['LNAME'], FILTER_SANITIZE_STRING);
-                    $subscriber->setSubscriberLastname($subscriberLname);
-                }
-                $this->subscribeMember($subscriber);
             }
         } catch (Exception $e) {
             Mage::logException($e);
@@ -171,7 +175,7 @@ class Ebizmarts_MailChimp_Model_ProcessWebhook
     protected function _unsubscribe(array $data)
     {
         $subscriber = Mage::helper('mailchimp')->loadListSubscriber($data['list_id'], $data['email']);
-        if ($subscriber->getId()) {
+        if ($subscriber && $subscriber->getId()) {
             try {
                 $action = isset($data['action']) ? $data['action'] : 'delete';
                 switch ($action) {
